Moved azureClients and types under azure folder
