+++ src/vscode-bicep/src/azure/azureClients.ts
