+++ src/vscode-bicep/src/azure/types.ts
